% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse spectra.R
\name{fuse_spectra}
\alias{fuse_spectra}
\title{Title Vector sum}
\usage{
fuse_spectra(x)
}
\arguments{
\item{input}{type of x is a matrix}
}
\value{
s is a vector
}
\description{
Title Vector sum
}
\examples{
data(iRaman)
x <- cbind(y1,y2,y3,y4,y5,y6)
s <- fuse_spectra(x)
}
