% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSVPdtw.R
\name{RSVPdtw}
\alias{RSVPdtw}
\title{Title Process of Raman standardization}
\usage{
RSVPdtw(m, s)
}
\arguments{
\item{input}{type of m is a vector}

\item{n}{is the result of warping path}
}
\value{
q is result of warped
}
\description{
Title Process of Raman standardization
}
\examples{
data(F900)
m <-  c(1:1573)
q <- RSVPdtw(m,s)
s <- s$s1
x <- reference1[1:1500]
y <- query1[1:1500]
cor.test(x,y)
}
