% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warping path.R
\name{warpingpath}
\alias{warpingpath}
\title{Title warping path of Raman standardization}
\usage{
warpingpath(x, y, d, e, maxshift)
}
\arguments{
\item{input}{type of y is a vector}
}
\value{
s is the warping path
}
\description{
Title warping path of Raman standardization
}
\examples{
data(iRaman)
x = ref_in
pretreated <- x - airPLS(x,10e7,4,0.8)
data(iRaman)
x = ref_in_bc
y = query_in_bc
s <- warpingpath(x,y,8,6,8)
}
