% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSVPdtw.R
\name{RSVPdtw}
\alias{RSVPdtw}
\title{Title Process of Raman standardization}
\usage{
RSVPdtw(m, s)
}
\arguments{
\item{input}{type of m is a vector}

\item{n}{is the result of warping path}
}
\value{
q is result of warped
}
\description{
Title Process of Raman standardization
}
\examples{
data(iRaman)
rs_min=300
rs_max=3000
m <- query_rs
query_rs=query_rs[query_rs>rs_min & query_rs<rs_max]
q <- RSVPdtw(m,s)
query_in_bc=query_rs[query_rs>rs_min & query_rs<rs_max]
query_interp = approx(query_rs,query_in_bc15, q, yleft=0, yright=0)$y
x <- ref_in_bc15[1:1600]
y <- query_interp[1:1600]
cor.test(x,y)
}
